GraphObjAttributes.readFloatAttributeFromFile now is                    readAttributeFromFile
this method tries to construct a double from the first
attribute it encounters; if it succeeds, then all
subsequent attributes in this file are also forced
to Double.  (on subsequent failure, an exception is thrown).

if the first conversion to Double fails, all attributes in
this file are interpreted as strings


git-svn-id: http://chianti.ucsd.edu/svn@4699 0ecc0d97-ab19-0410-9704-bfe1a75892f5