added new method:
   public void cloneWindow

a new action

   protected class CloneGraphInNewWindowAction

and a new entry on the View menu.

this comment accompanies the cloneWindow method:

/**
 *  create a new CytoscapeWindow with an exact copy of the current window,
 *  nodes, edges, and all attributes.  <p>
 *  as a temporary expedient, this method piggybacks upon the already
 *  existing and tested 'SelectedSubGraphFactory' class, by first selecting all
 *  nodes in the current window and then invoking the factory.  when time
 *  permits, the two stages (select, clone) should be refactored into two
 *  independent operations,  probably by creating a GraphFactory class, which
 *  clones nodes, edges, and attributes.
 */

please note:  this version of CytoscapeWindow is the result of
a cvs merge of my version 1.45 and the archive's 1.52.  I do
not believe I did any harm in the merge and update, but if
you experience odd behavior, these changes are a good suspect
to consider.


git-svn-id: http://chianti.ucsd.edu/svn@4828 0ecc0d97-ab19-0410-9704-bfe1a75892f5