Added method isHeaderLineMTXHeader(String hline) which returns true if the given line is a header for a MTX file.
The file format that ExpressionData accepts is essentially the same as an MTX file, except that it does not have a header line of the form:
\t+RATIOS\t+LAMBDAS
It is convenient for users to be able to input their MTX files directly, whitout having to remember to manually remove the first line of their files. Therefore, I made a small modification to the loadData method so that both file formats are accepted (MTX, or MTX without first line).

-iliana


git-svn-id: http://chianti.ucsd.edu/svn@5087 0ecc0d97-ab19-0410-9704-bfe1a75892f5