owen's change from last week:
  I noticed that the AnnotationGui was leaving String[]'s in
  nodeAttributes, rather than using the GraphObjAttributes method "append"
  and using the Vectors that result from that.  So I changed it to use
  Vectors, assuming that the GraphObjAttributes interface is to be
  respected.

broke a few things in cytoscape.data.annotation.AnnotationGui and
                      cytoscape.layouts.AttributeLayout

which are now fixed.  The downstream clients of that annotation expected
String [], and got Vectors of Strings.  Since Owen's approach is indeed
the better way to go, I fixed the downstream clients.

I also moved 'unpackPossiblyCompoundStringAttributeValue' -- which had
existed as two identical copies in the above-mentioned classes -- into
a static method in GraphObjAttributes.


git-svn-id: http://chianti.ucsd.edu/svn@5145 0ecc0d97-ab19-0410-9704-bfe1a75892f5