Added DefaultPanel class to the makefile.
Fixed a bug in DefaultPanel in which setting the background color
didn't work once a graph was loaded. Tracked this down to the fact
that CytoscapeWindow keeps it's own GraphView, which is shown in the
window but is not the current view of the Graph. Fixed this by changing
the constructor of DefaultPanel to take a VisualMappingManager object,
and when it wants to get or set the color it does a dynamic grab of
the view via vmm.getCytoscapeWindow().getGraphView(). Verified that
this fixes the bug.
-Andrew 03-14-03


git-svn-id: http://chianti.ucsd.edu/svn@5246 0ecc0d97-ab19-0410-9704-bfe1a75892f5