the ctor now calls String.trim () on every token it parses out of the rawText (which parsing is done on <delimiter> boundaries).
this changes nothing when the delimiter is a space, but if there
are spaces at the beginning or end of tokens, and the
term-interaction-term delimiter is a tab, these spaces are now
stripped off.

this change is occasioned by a user who used tabs to delimit
interaction terms, and left a trailing space; two apparently
identical protein names ended up as different names, due
to the one trailing space.


git-svn-id: http://chianti.ucsd.edu/svn@5355 0ecc0d97-ab19-0410-9704-bfe1a75892f5