the server can now consist of a URL, in which case the HttpTextReader will be used.
it used to be that all files were assumed to be xml unless they
ended in 'txt'.  no the sense of that is reversed:  all files are
assumed to be text files unless they end in xml.  (and note that
using xml is pretty much abandoned:  using my approach to parsing,
it just took too long to build a dom document; reading text
is much much quicker)


git-svn-id: http://chianti.ucsd.edu/svn@5388 0ecc0d97-ab19-0410-9704-bfe1a75892f5