Modifications throughout the visual package to be consistent with the refactored core.
Changed all uses of cytoscape.visual.Network to cytoscape.data.CyNetwork.
Rewrote the main classes to use an instance of NetworkView instead of
references to CytoscapeWindow. The changed intriduced a new method
called "setUI" in VisualMappingManager, so that this class can inform
the UI when the visual style is changed. As mentioned in the javadoc,
there are better ways to hand

Changed CyWindow to use different constructors for the visual mapper
and the UI.
-Andrew 08-22-03


git-svn-id: http://chianti.ucsd.edu/svn@5418 0ecc0d97-ab19-0410-9704-bfe1a75892f5