Checking in major refactorings of the action classes and some dialogs assoicated with menu items in Cytoscape.
In the actions/dialogs replaced CytoscapeWindow references with references
either to CyWindow, or more commonly to the NetworkView interface
that CyWindow implements.

Moved to archive directories classes that are not currently referenced
by anything else in Cytoscape, rather than editing them.

Added new class Semantics to handle naming issues and the species
attribute as helpers to the rest of Cytoscape. Added methods to
CyNetworkUtilities that replace matching methods in CytoscapeWindow,
in most cases redirecting the CytoscapeWindow methods to these
new methods. Rewrote several actions to refer to these classes.

Edited CytoscapeMenus to match the new constructors of the actions.
-Andrew 09-26-03


git-svn-id: http://chianti.ucsd.edu/svn@5446 0ecc0d97-ab19-0410-9704-bfe1a75892f5