Rewrote Cytoscape's plugin loading system.
The actual plugin instance construction is now in a static method of
AbstractPlugin. This allows us to easily extend the set of constructors
we recognize. Also, this method was extended to recognize plugins with
a CyWindow argument constructor in addition to the CytoscapeWindow
constructor.

PluginLoader and JarClassLoader were rewritten to delegate to this new
method of AbstractPlugin.

All of the plugin loading classes were converted from CytoscapeWindow
references to CyWindow references.

I had to change the CytoscapeWindow constructor that is called from
CyWindow. If we try to do plugin loading directly, then CyWindow
doesn't have a valid CytoscapeWindow reference to give to the plugin
loader. So, for only this CytoscapeWindow constructor, we don't call
loadPlugins() directly but rather have CyWindow call loadPlugins()
after calling the constructor.

If there are no menu options from plugins, then we provide a default
non-operational menu filler for the operations menu. I moved the logic
of checking for this case to CyMenus and rewrote everyone else to
delegate to this method after doing plugin loading.

-Andrew 11-10-03


git-svn-id: http://chianti.ucsd.edu/svn@5508 0ecc0d97-ab19-0410-9704-bfe1a75892f5