class moved from package cytoscape.jarLoader. Old Log follows: ---------------------------------------------------------------------- RCS file: /common/cvsdir5/cytoscape/src/cytoscape/jarLoader/JarClassLoader.java,v Working file: JarClassLoader.java head: 1.9 branch: locks: strict access list: symbolic names: 	yFiles-partial-removal: 1.4 	cyto-2_0-alpha1: 1.4 	public-1_1: 1.3 keyword substitution: kv total revisions: 9;	selected revisions: 9 description: ---------------------------- revision 1.9 date: 2004/04/06 04:31:15;  author: sheridan;  state: Exp;  lines: +6 -6 Functions for registering plugins encapsulated into new class: PluginRegistry ---------------------------- revision 1.8 date: 2004/04/06 02:44:57;  author: sheridan;  state: Exp;  lines: +154 -164 Update to plugin loading architecture. Various java doc comments updated in afffected files.
The purpose of these changes was to centralize the plugin loading
system so that plugins loaded in one window would be available in
other windows. Plugins are still attached to a CyWindow, but instead
of creating them directly there, plugins are instead loaded into
a registry (currently centralized in CytoscapeObj. This registry is
shared across an entire running instance of Cytoscape. Each window
will check its own plugin menu whenever the plugin menu is opened.
If there have been changes to the central registry since the last
time this window's menu has been updated, the menus is again updated
to include any plugins which have been loaded since the last update.
Missing plugins are instanciated with a link to the current CyWindow.

This plugin registry function will immediately be seperated out into
a new PluginRegistry class .. in an update to be made later today.

Other changes include:
- Dropping the support for conditional loading of plugins from the
cytoscape.props properties file. This allowed the code to be greatly
simplified, and that feature was not currently working correctly.
There are now 3 ways to load plugins: Through the user menu; with
the command line argument --JLD; and with an entry in the cytoscape.props
file of the form plugin.<name of plugin>.load<name of plugin>.

Changes were made to the startup scripts to allow the property file
loading method to work. Previously, these scripts used a command
like "java -jar cytoscape.jar" and this form of starting the VM
will *ignore* any CLASSPATH setting or -classpath command line
argument. Therefore, any class name which was included in the
properties file was not able to be found by the system classloader.
Instead, the startup scripts now do something like:
"java -classpath 'cytoscape.jar:.:$CLASSPATH' cytoscape.CyMain"
to start cytoscape. If CLASSPATH is set correctly, plugins can
now be loaded through the properties file.

Other functions have been consolidated in CyMenus. This is in
preparation for moving classes to a newly planned subpackage:
"cytoscape.plugin"
----------------------------
revision 1.7
date: 2004/03/25 22:35:07;  author: sheridan;  state: Exp;  lines: +7 -12
New windows are populated with all loaded Jar plugins
Loading a plugin Jar causes it to be added to all existing windows
- CytoscapeObj now modified to include a registry of loaded plugins
	(this probably should be moved to a seperate class)
- PluginLoader and JarClassLoader: loaded plugin classes are
	added to the registry rather than being immediately
	loaded into the current window.
- CyMenus.initializeMenus() now adds a MenuListener instance to
	the opsMenu: when the menu is opened, the plugins loaded
	since the last time this menu was opened are instantiated
	for this window and added to the menu.
----------------------------
revision 1.6
date: 2004/03/22 23:11:45;  author: sheridan;  state: Exp;  lines: +40 -51
Plugin loader now tracks loaded plugins and rejects classes which
	have previously been loaded as plugins. (see bug 187)
----------------------------
revision 1.5
date: 2004/02/20 01:25:31;  author: xmas;  state: Exp;  lines: +139 -118
Numerous Changes, mostly with some menu stuff.
----------------------------
revision 1.4
date: 2003/11/10 22:02:47;  author: amarkiel;  state: Exp;  lines: +13 -32

Rewrote Cytoscape's plugin loading system.

The actual plugin instance construction is now in a static method of
AbstractPlugin. This allows us to easily extend the set of constructors
we recognize. Also, this method was extended to recognize plugins with
a CyWindow argument constructor in addition to the CytoscapeWindow
constructor.

PluginLoader and JarClassLoader were rewritten to delegate to this new
method of AbstractPlugin.

All of the plugin loading classes were converted from CytoscapeWindow
references to CyWindow references.

I had to change the CytoscapeWindow constructor that is called from
CyWindow. If we try to do plugin loading directly, then CyWindow
doesn't have a valid CytoscapeWindow reference to give to the plugin
loader. So, for only this CytoscapeWindow constructor, we don't call
loadPlugins() directly but rather have CyWindow call loadPlugins()
after calling the constructor.

If there are no menu options from plugins, then we provide a default
non-operational menu filler for the operations menu. I moved the logic
of checking for this case to CyMenus and rewrote everyone else to
delegate to this method after doing plugin loading.

-Andrew 11-10-03
----------------------------
revision 1.3
date: 2003/05/06 19:50:20;  author: owen;  state: Exp;  lines: +18 -2
Now the plugins menu is not emptied when loadPlugins() is called,
which makes it possible for a subsequent JarLoader call for command
line arguments to succeed.  Also, JarLoader knows to remove the
"No plugins loaded" message if necessary; finally, the makefile has
been updated to depend on the jarLoader directory.
----------------------------
revision 1.2
date: 2003/04/24 15:46:51;  author: owen;  state: Exp;  lines: +2 -2
included package cytoscape; now compiles.
----------------------------
revision 1.1
date: 2003/04/23 22:15:45;  author: owen;  state: Exp;
began adding jarLoader from csplugins to cytoscape's core.
=============================================================================


git-svn-id: http://chianti.ucsd.edu/svn@5761 0ecc0d97-ab19-0410-9704-bfe1a75892f5