You should notice these tags on all files:
  pre-rollback-of-2005_04_22-changes
  post-2005_04_22-destructive-changes
  pre-2005_04_22-destructive-changes

pre-rollback-of-2005_04_22-changes is essentially the exact state of CVS
at the time this commit was made.  Consider it so.

What I've done is taken the CVS state at snapshot
pre-rollback-of-2005_04_22-changes, then applied the diffs going from
post-2005_04_22-destructive-changes to
pre-2005_04_22-destructive-changes to that snapshot.  I then ironed out the
compile-time errors.  This is what I'm committing.  My motivation
for this huge change is to re-add valuable code such as the biodataserver.
Also, the April 22 destructive chages included numerous incompatiblle public
API changes.  We don't make such incompatible API changes.


git-svn-id: http://chianti.ucsd.edu/svn@6563 0ecc0d97-ab19-0410-9704-bfe1a75892f5