edge-directionality branch: honor directedness flag in Cytoscape.getCyEdge()
This fixes the previously failing tests in CytoscapeTest.java.

Also set the version of .getCyEdge() that doesn't have a directedness
arg to create directed edges. This contradicts previous javadoc of that method,
but that was a bug in the comment: all edges were created as directed
(as that is used as default in RootGraph.java).

Update javadoc comment accordingly.


git-svn-id: http://chianti.ucsd.edu/svn@14083 0ecc0d97-ab19-0410-9704-bfe1a75892f5