edge-directionality branch: don't draw arrows or edge ends for undirected edges
As the two ends of and undirected edge are supposed to be equivalent, 
drawing them with different edge-ends would be highly misleading. As
allowing some way to set the edge-end of undirected edges (i.e. both to the
same setting) would be complicated, the simplest fix is to force
both ends of undirected edges to be NO_END (i.e. no arrows).



git-svn-id: http://chianti.ucsd.edu/svn@14314 0ecc0d97-ab19-0410-9704-bfe1a75892f5