pluggable-renderers: define org.cytoscape.view.VisualProperty and org.cytoscape.view.VisualPropertyCatalog; copy all code from org.cytoscape.vizmap.properties to cytoscape.LegacyVisualProperty which will be used as a placeholder during the refactoring (and then removed).
Populate VisualPropertyCatalog with legacy VisualProperties during
initialization of Cytoscape.

(Note that the code that is replaced by all this is not yet removed,
because it is used in many places and a lot of stuff breaks. This means
that current code is not actually used until the next step in refactoring.)


git-svn-id: http://chianti.ucsd.edu/svn@14587 0ecc0d97-ab19-0410-9704-bfe1a75892f5