pluggable-renderers: turn inside loop of applying VisualStyles inside-out and move it from AbstractCalculator to VisualStyle.
This is needed because previous version applied the VisualStyle by applying the Calculators
one-by-one, which meant that if a Calculator was not defined for a given VisualProperty,
VizualMappingBypass was not applied either. 

(in pre-refactoring code, the VisualMappingBypass was applied seperately on the NodeAppearance,
EdgeAppearance objects, thus such a change wasn't needed).


git-svn-id: http://chianti.ucsd.edu/svn@14634 0ecc0d97-ab19-0410-9704-bfe1a75892f5