pluggable-renderers: remove performance-hack from VizMapperMainPanel: remove 'ignore' flag and VizMapperMainPanel.enableListeners()
The purpose of these where to allow VizMapperMainPanel ignore events
when opening a session since a lot of events would be generated.  I am
removing this hack because it hinders refactoring and debugging, and
removing it will remove a dependency of
cytoscape.data.readers.CytoscapeSessionReade on VizMapperMainPanel
(which dependency breaks modularity), and also because such
performance-optimization should be implemented in the event-firing
part of the code, not in the event-listener part. It should be a
general part of the event-firing API, not a one-off hack.


git-svn-id: http://chianti.ucsd.edu/svn@15162 0ecc0d97-ab19-0410-9704-bfe1a75892f5