pluggable-renderers: sprinkle some refresh calls and event handlers, in the hope of fixing some refresh bugs.
Note that most likely this is not the real solution (even though it
appears to make the situation better). The event handling and
propagation of the vizmapper gui should be re-thought and ground-up
redesigned / audited. For example, VisualPropertySheetPanel can't do a
refresh in a CytoscapeDesktop.NETWORK_VIEW_FOCUS or
NETWORK_VIEW_FOCUSED event callback, since it will rely on the
VizMapperMainPanel already having adapted to the currentView state,
but that might depend on which order these event callbacks are called,
which might be a bad thing to depend on. Instead, the
VizMapperMainPanel should have (and fire) events like
CURRENTLY_EDITED_VS_CHANGED and CURRENTLY_EDITED_VIEW_CHANGED, and
these sub-widgets should listen to that. Since I wouldn't be able to
finish that by the second mini-retreat (this Friday), as a work-around
for now sprinkle some extra event-handling code. For example, call
visualPropertySheetPanel.adaptToVisualStyleChanged(); from
VizMapperMainPanel.adaptToVisualStyleChanged().



git-svn-id: http://chianti.ucsd.edu/svn@15245 0ecc0d97-ab19-0410-9704-bfe1a75892f5