pluggable-renderers: some hacks to fix some refresh bugs.
Note that this is just to get the code working for the mini-retreat,
these changes most likely introduce more problems than they fix, and
they don't fix the underlying problems. For example,
DefaultAppearenceBuilder and DefaultViewPanel need to be completely
redesigned, but I won't have time to do that before the mini-retreat.


git-svn-id: http://chianti.ucsd.edu/svn@15250 0ecc0d97-ab19-0410-9704-bfe1a75892f5