refactored-viewmodel: implement MappingCalculator that uses column-oriented API This requires changes to public api: - make ViewColumn<T> part of public API - change signature of MappingCalculator.apply() to operate on a column   instead of on a single View - make MappingCalculator a non-generic type, ie. remove T   type-parameter - in integration-test: add Integer-to-Number and Double-to-Number   mapping examples. - add run-column-oriented.sh to compile and run column-oriented branch
As far as I know, this version has no type-safety warnings, apart from
those due to lack of 'type-parameter preserving map'. Since
MappingCalculator has no type-parameter, this version demonstrates a
usable option for implementing a generic and type-safe
PassthroughMappingCalculator.

Note that this version is not actually type-safe, since the possible
types that PassthroughMappingCalculator can handle are not stored or
checked anywhere. In case of type mismatch,
PassthroughMappingCalculator will silently do nothing. (This is
demonstrated in CyMain.java)



git-svn-id: http://chianti.ucsd.edu/svn@15536 0ecc0d97-ab19-0410-9704-bfe1a75892f5