refactored-viewmodel: make MappingCalculator non-generic in row-oriented implementation / API
This was pretty simple, the cost is only one
static-type-safety-breaking cast in
PassthroughMappingCalculator.apply, which can be shown to be type-safe
due to other constraints.

Also, a small fix, which changes API:
since MappingCalculator implementations will skip those rows where the
attribute being mapped from is not defined, they will have to apply
the default value in these cases. Thus MappingCalculator.apply() needs
to accept the value as an extra arg.



git-svn-id: http://chianti.ucsd.edu/svn@15539 0ecc0d97-ab19-0410-9704-bfe1a75892f5