refactored-viewmodel: in column-oriented implementation: clear old values when applying VisualStyle
This is needed so that the contents of the viewmodel won't keep values
that were set by previous VisualStyle even if the new VisualStyle does
not define values for them. I.e. to revert these 'cells' to the
default values.

This is done automatically in the row-oriented implementation, since
there the default value is not stored in the column, but in the cell
for the View.

API change: needed to add a ViewColumn.clearValue() method, to allow
'erasing a cell'


git-svn-id: http://chianti.ucsd.edu/svn@15541 0ecc0d97-ab19-0410-9704-bfe1a75892f5