refactored-viewmodel: branch vizmap.mappings api that is used in vizmap-gui from cytoscape3 trunk Copy over .java files from under svn+ssh://grenache.ucsd.edu/cellar/common/svn/cytoscape3/trunk/vizmap/src/main/java/org/cytoscape/vizmap/mappings/
Note that ContinuousMappingPoint.java was copied from
vizmap.mappings.continuous but since it uses stuff in vizmap.mappings
and itself is used in vizmap.mappings it migh as well be in
vizmap.mappings.

Only part of api that vizmap-gui uses and is not copied now is
org.cytoscape.vizmap.mappings.ObjectMapping since that api is covered
by MappingCalculator

No local changes made yet, so this will not compile at all.


git-svn-id: http://chianti.ucsd.edu/svn@15563 0ecc0d97-ab19-0410-9704-bfe1a75892f5