Version 0.71: Data display has been added to the main class for debug; it retains all the previous functionality. GeometryUtil: Added a control point calculator (CP) to find CP' from original CP, such that when curveTo accepts CP' as input, the curve is forced to pass through CP.
Version 0.72: 
Added functionality to set different gridResolution for GBEB. Notes: I didnt not add a functionality for gridSize adjustments as 
such a function would incur a huge memory cost as the whole operator has to be cloned in order to save on computational speed.
Debugging the problem which causes excessively long edges. 
GBEBRouter: - MergeShape: All grid Index is now unique
						- Known issue 1 is fixed
KDE: If there is < 2 edges, it would not be strongly clustered.

Version 0.73:
GBEBRouter: - tweaked the mergeNodes_All() such that it is consistent with new GBEBRouter class, also modified code such
that when merging the edges extends but do not move away from control points. 
						- Added a function (displayUnmergedMesh) to display the original mesh
						- Changed mergeShape to also merge neighbouring Shapes: Decreased the occurance of "Ladders" - known Issue 6. - but
								this in turn creates a situation where the mesh edge might not lie within the Shape.
						- fixed problem of meshEdge intersecting with Edges that are in the shape but does not agree with general directions
						- fixed curve makeing loops
						- adjusted the quadratic curve for eqn for EdgeRenderer such that it passes through the CP

git-svn-id: http://chianti.ucsd.edu/svn@21079 0ecc0d97-ab19-0410-9704-bfe1a75892f5