A new package (ivis) is added under org/cytoscapeweb/view/layout package. This package includes the layout and utility classes required to run CoSE within CytoscapeWeb. CoSE (Compound Spring Embedder) layout is a part of Chisio project, and most of the classes contained in ivis package are produced by converting required parts of Chisio from Java to AS3. For collection classes, as3commons-collection library added under the lib directory.
A new class (ForceDirectedCompoundLayout) is added under org/cytoscapeweb/view/layout package to create the graph topology for CoSE and to update the node sprite positions after applying the layout operations. Also, classes related to layout options and operations are modified in order to integrate the new layout to CytoscapeWeb.

On the other hand, a minor bug related to default visual style naming for compound nodes is fixed within the VisualStyleVO class.

git-svn-id: http://chianti.ucsd.edu/svn@24837 0ecc0d97-ab19-0410-9704-bfe1a75892f5