fixes #848 the code for the the initial indexing was commented out for performance issues. It may cause some performance hits depending on the number and size of networks and the threads running. Currently since the handling event for current network change is invoked in another thread it does not cause the other parts of application to wait for the initiation.
The handleAttributesChanged is also called in the network focus changed handler to update the attributes combobox based on the current network for adding different widgets. The column name changed listener is also extended and handled to update the attributes combobox if a column's name is changed. However, it is possible to have a widget already added for the renamed column in this case the filter will not be effective since the column does not exist anymore.



git-svn-id: http://chianti.ucsd.edu/svn@28811 0ecc0d97-ab19-0410-9704-bfe1a75892f5